Move and add view logics outside fragment into their own view holder class

+++ features/merchant/seller_home/src/main/java/com/tokopedia/sellerhome/settings/view/adapter/uimodel/RmTransactionData.kt
@@ -1,7 +1,7 @@
 package com.tokopedia.sellerhome.settings.view.adapter.uimodel
 
 data class RmTransactionData(
-    val totalTransaction: Int,
+    val totalTransaction: Long,
     val dateCreated: String,
     val isBeforeOnDate: Boolean
 )
